+++ src/questions/questions.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { PrismaService } from 'src/database/prisma.service';
 import { CreateQuestionsDTO } from './dtos/create.questions.tdo';
 import { IQuestions } from './interfaces/questions.interface';
@@ -9,42 +9,51 @@ export class QuestionsService {
 
     constructor (private prisma: PrismaService) {}
 
-    create(question:CreateQuestionsDTO): Promise<IQuestions>{
-        return this.prisma.questions.create({data:question})
+    async create(question:CreateQuestionsDTO): Promise<IQuestions>{
+        return await this.prisma.questions.create({data: question});  
     }
 
-    findAll(): Promise<IQuestions[]>{
-        return this.prisma.questions.findMany()
+    async findAll(): Promise<IQuestions[]>{
+        return await this.prisma.questions.findMany()
     }
 
-    async findById(id:number): Promise<IQuestions> {
-        let numberId = parseInt(id.toString())
-        const question = await this.prisma.questions.findUnique({
-            where: {id:numberId}
-        })
-        if(question) {
-            return question
-        }
-        throw new NotFoundException('Questão não encontrada');
-    }
-
-    async update(id:number, question:UpdateQuestionsDto): Promise<IQuestions>{
+    async findByID(id: number): Promise<IQuestions> {
         let number = parseInt(id.toString())
-        try {
-            const updatedQuestion = await this.prisma.questions.update({
-                where: { id:number },
-                data: question,
-            });
-            return updatedQuestion;
-        } catch (error) {
-            throw new NotFoundException('Questão não encontrada');
+        const user = await this.prisma.questions.findUnique({
+            where: { id: number }
+        });
+        
+        if (user) {
+            return user
         }
+    }
+
+    async update(id: number, users: UpdateQuestionsDto, request): Promise<IQuestions>{
+        if (!request.user) {
+            throw new UnauthorizedException("Usuário não autenticado");
+          }
+        let number = parseInt(id.toString())
+        if (request.user.id === number || request.user.admin === true){
+            try {
+                const updatedUser = await this.prisma.questions.update({
+                    where: { id:number },
+                    data: users,
+                });
+                return updatedUser;
+            } catch (error) {
+                throw new NotFoundException('Usuário não encontrado');
+            }
+        }
+        throw new UnauthorizedException("Seu usuario não pode editar")
     }
     
-    delete(id:number): Promise<IQuestions>{
+    async delete(id: number, request): Promise<IQuestions>{
         let number = parseInt(id.toString())
-        return this.prisma.questions.delete({
-            where: {id: number}
-        })
+        if (request.user.admin === true || request.user.id === number) {    
+            return this.prisma.questions.delete({
+                where: {id: number},
+            });
+        }
+        throw new UnauthorizedException("Seu usuario não pode excluir tarefas")
     }
 }
